Make running in container easier, by supporting envvars in config
This adds basic, but pretty safe, support for interpolation of envvars into the
main and collector configuration files.

This is to make running Fullerite inside a container on a 12-factor-like PaaS
(eg Yelp/Paasta) more straightforward.

It works as a simple templating engine, replacing strings like %%MY_ENVVAR%%
with their corresponding value.

Specifically this works by:

* Looking for strings matching %%[A-Z][A-Z0-9_]{4,}%%
   - minimum 5 chars
   - must be upper case + numeric + underscore
   - must start with letter
* For each string matching that, pull out the envvar name inside the %%'s
* If there is an envvar matching that name, substitute it in
* If there is no envvar matching that name, leave it as-is.

This is designed to be as safe as possible to existing values, and high entropy
values like authentication tokens (where we could get a collision with
%%SOMETHING%%.

Arguably we could use a proper templating engine like Mustache for this, but
this is simple and effective.